huge cleanup, i18n initial support
