+++ package.json
@@ -20,6 +20,7 @@
   },
   "dependencies": {
     "@arrow-js/core": "latest",
-    "lisan": "latest"
+    "lisan": "latest",
+    "mousetrap": "latest"
   }
 }
