+++ src/languages/en-US.js
@@ -1,0 +1,52 @@
+module.exports = {
+	locale: 'en-US',
+	keys: {
+		"prompt.countOptions": options => `Count ${options.length}`,
+		"prompt.getSettings": "Get Settings",
+		"prompt.saveSettings": "Save Settings",
+		"prompt.getCommands": "Get Commands",
+		"menu.explanation": `Settings stuff`,
+		"settings.theme": "Theme",
+		"settings.useApiName": "Show API Name",
+		"commands.home": "Home",
+		"commands.setup": "Setup",
+		"commands.mergeAccounts": "Merge Accounts",
+		"commands.toggleAllCheckboxes": "Toggle All Checkboxes",
+		"commands.toggleLightning": "Toggle Lightning",
+		"commands.loginAs": "Login As [login as <FirstName> <LastName> OR <Username>]",
+		"commands.search": "?",
+		"commands.objectManager": "Object Manager",
+		"commands.toggleEnhancedProfiles": "Toggle Enhanced Profiles",
+		"commands.refreshMetadata": "⚙️ Refresh Metadata",
+		"commands.dumpDebug": "⚙️ Dump Debug Info to Console",
+		"commands.setSearchLimit": "⚙️ Set Search Limit [set search result limit <##> (where the <##> is 32 or less)]",
+		"commands.themesDefault": "⚙️ Set Theme: Default",
+		"commands.themesDark": "⚙️ Set Theme: Dark",
+		"commands.themesUnicorn": "⚙️ Set Theme: Unicorn",
+		"commands.themesSolarized": "⚙️ Set Theme: Solarized",
+		"commands.errorAccountMerge": "[ERROR: you must have an Account ID in your clipboard or type one after the merge command]",
+		"commands.goToTask": "Go To Created Task",
+		"commands.escapeCommand": "(press escape to exit or enter a new command)",
+		"notification.searchSettingsUpdated": "Search settings updated",
+		"error.searchLimitMax": "Be sure to set your search limit to 32 or less",
+		"error.notFound": " not found in command list or incompatible",
+		"prefix.list": "List ",
+		"prefix.new": "New ",
+		"prefix.setup": "Setup > ",
+		"prefix.loginAs": "Login As ",
+		"prefix.error": "Error! ",
+		"objects.fieldsAndRelationships": "Fields",
+		"objects.pageLayouts": "Page Layouts",
+		"objects.lightningPages": "Lightning Record Pages",
+		"objects.buttonsLinksActions": "Buttons, Links, and Actions",
+		"objects.compactLayouts": "Compact Layouts",
+		"objects.fieldSets": "Field Sets",
+		"objects.limits": "Limits",
+		"objects.recordTypes": "Record Types",
+		"objects.relatedLookupFilters": "Related Lookup Filters",
+		"objects.searchLayouts": "Search Layouts",
+		"objects.triggers": "Triggers",
+		"objects.lightningPages": "Lightning Record Pages",
+		"objects.validationRules": "Validation Rules"
+	},
+}
