+++ src/main.js
@@ -1,0 +1,15 @@
+// @copyright 2012+ Daniel Nakov / Silverline CRM
+// http://silverlinecrm.com
+// @copyright 2019+ Danny Summerlin
+
+var sfnav = (()=>{
+
+// login as
+
+
+// interaction
+
+
+
+// setup
+	function ()
